requirements.txt
aio-pika==9.4.0
aiormq==6.8.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
cached-property==1.5.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
contourpy==1.0.7
cryptography==42.0.4
cycler==0.11.0
deprecation==2.1.0
distlib==0.3.8
docker==6.0.1
ecdsa==0.18.0
fastapi==0.105.0
filelock==3.13.1
fonttools==4.39.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.4
iniconfig==2.0.0
joblib==1.3.2
keycloak-client==0.15.4
kiwisolver==1.4.4
llvmlite==0.41.1
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.7.1
multidict==6.0.5
numba==0.58.1
numpy==1.24.2
packaging==23.0
pamqp==3.3.0
pandas==2.1.0
pika==1.3.2
Pillow==9.4.0
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.4.0
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.20.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.2.0
pydantic_core==2.14.5
pynndescent==0.5.11
pyparsing==3.0.9
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==3.9.0
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.2
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.27.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
umap==0.1.1
umap-learn==0.5.5
urllib3==1.26.15
uvicorn==0.24.0.post1
virtualenv==20.25.0
vk==3.0
vk-api==11.9.9
websocket-client==1.5.1
yarl==1.9.4
